Update baby_agi.ipynb
Add a sample description to use ElasticSearch V8.8.1 as the vectorstore. There's something to be told about how to initialize the vectorstore with proper metadatas. As currently the BabyAGI code is hardcoded to read the ['task'] metadata key. I spent a lot time to find out the reason and fixed my test run. Please review. Hope this will help others who want to run this great demo using ElasticSearch.